Deploy python package to PyPi
This PR moves the python package to the 'src/package' organization
standard, as well as adds a github action step to package and deploy the
library to PyPi.  Because of this, some additional modifications were
made to the pyproject.toml in order to provide more complete descriptions,
as well as support setuptools-scm for easier distribution purposes.
Finally the .gitignore file was updated to exclude some of the python
packaging files that are created.